feat: create new EcosystemPartnersTable component that uses the headless EcosystemPartners component
