+++ components/content/EcosystemPartners.vue
@@ -1,14 +1,16 @@
 <script setup lang="ts">
   import { PropType } from 'vue';
 
-  import { EcosystemConfig, EcosystemPartner } from '@/config/ecosystem';
   import { useExternalJSON } from '@/data/useExternalJSON';
+  import { EcosystemPartners } from '@/domain';
+
+  import { EcosystemPartnerType } from '@/types';
 
   const props = defineProps({
-    type: { type: String as PropType<EcosystemPartner>, required: true },
+    type: { type: String as PropType<EcosystemPartnerType>, required: true },
   });
 
-  const { data, isLoading } = await useExternalJSON(EcosystemConfig[props.type]);
+  const { data, isLoading } = await useExternalJSON(EcosystemPartners.getJSONFileUrls(props.type));
 </script>
 
 <template>
