+++ content/5.resources/2.ibc-channels.md
@@ -4,34 +4,26 @@ title: IBC channels
 description: This page serves as a reference for anyone looking to explore the various IBC channels that connect the Archway network to other networks.
 parentSection: Resources
 parentSectionPath: /resources
-testnetRelayerFiles: ['constantine-3-osmosistestnet.json','constantine-3-axelartestnet.json','constantine-3-cosmoshubtestnet.json']
-mainnetRelayerFiles: ['archway-osmosis.json', 'archway-axelar.json','archway-cosmoshub.json',]
-
 ---
 
 # IBC Channels
 
 Here, you can find a comprehensive list of channels that connect the Archway blockchain network with other IBC-enabled networks.
 
-
 ## Triomphe (Mainnet)
 
 The chain ID for the Triomphe mainnet is `archway-1`.
 
-::relayersTable{:repoFiles="mainnetRelayerFiles"}
+::ecosystem-partners-table{type="mainnetRelayers"}
 ::
-<!-- <relayers-table :repo-files="mainnetRelayerFiles" /> -->
 
 You can also find the list of IBC channels in our github <a href="https://github.com/archway-network/networks/tree/main/_IBC" target="_blank" >networks</a> repo.
-
-
 
 ## Constantine (Testnet)
 
 The chain ID for the Constantine testnet is `constantine-3`.
 
-::relayersTable{:repoFiles="testnetRelayerFiles"}
+::ecosystem-partners-table{type="testnetRelayers"}
 ::
-<!-- <relayers-table :repo-files="testnetRelayerFiles" /> -->
 
 You can also find the list of IBC channels in our github <a href="https://github.com/archway-network/networks/tree/main/_IBC" target="_blank" >networks</a> repo.
