+++ types/EcosystemPartner.ts
@@ -1,0 +1,4 @@
+export enum EcosystemPartnerType {
+  MainnetRelayers = 'mainnetRelayers',
+  TestnetRelayers = 'testnetRelayers',
+}
