+++ types/index.ts
@@ -1,6 +1,7 @@
 import { AutocompleteComponents, HTMLTemplate } from '@algolia/autocomplete-js';
 import { BaseItem } from '@algolia/autocomplete-core';
 
+export * from './EcosystemPartner';
 export * from './IBCRelayer';
 
 export type AppConfigFeatures = {
